refactor: simplify db config and knexfile management
